[Minor] Types for Tabs component
